// src/gql/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** Big number integer */
  BigInt: any;
  /** A date-time string in simplified extended ISO 8601 format (YYYY-MM-DDTHH:mm:ss.sssZ) */
  DateTime: any;
};
// ...
export type Channel = {
  __typename?: 'Channel';
  /** Channel's avatar photo asset. */
  avatarPhoto?: Maybe<StorageDataObject>;
  /** List of members blocked from commenting/reacting on any video of the channel. */
  bannedMembers: Array<BannedMember>;
  /** Value of channel state bloat bond fee paid by channel creator */
  channelStateBloatBond: Scalars['BigInt'];
  /** Channel's cover (background) photo asset. Recommended ratio: 16:9. */
  coverPhoto?: Maybe<StorageDataObject>;
  /** Timestamp of the block the channel was created at */
  createdAt: Scalars['DateTime'];
  /** Number of the block the channel was created in */
  createdInBlock: Scalars['Int'];
  /** Cumulative rewards claimed by this channel */
  cumulativeRewardClaimed?: Maybe<Scalars['BigInt']>;
  /** The description of a Channel */
  description?: Maybe<Scalars['String']>;
  /** Application used for channel creation */
  entryApp?: Maybe<App>;
  /** Number of active follows (to speed up orderBy queries by avoiding COUNT aggregation) */
  followsNum: Scalars['Int'];
  /** Runtime entity identifier (EntityId) */
  id: Scalars['String'];
  /** Flag signaling whether a channel is censored. */
  isCensored: Scalars['Boolean'];
  /** Whether a channel has been excluded/hidden (by the gateway operator) */
  isExcluded: Scalars['Boolean'];
  /** Flag signaling whether a channel is public. */
  isPublic?: Maybe<Scalars['Boolean']>;
  /** The primary langauge of the channel's content */
  language?: Maybe<Scalars['String']>;
  /** Current member-owner of the channel (if owned by a member) */
  ownerMember?: Maybe<Membership>;
  /** Channel's reward account, storing the income from the nft sales and channel payouts. */
  rewardAccount: Scalars['String'];
  /** The title of the Channel */
  title?: Maybe<Scalars['String']>;
  /** Number of videos ever created in this channel */
  totalVideosCreated: Scalars['Int'];
  /** Number of total video views (to speed up orderBy queries by avoiding COUNT aggregation) */
  videoViewsNum: Scalars['Int'];
  /** List of videos that belong to the channel */
  videos: Array<Video>;
};
// ...
export type Comment = {
  __typename?: 'Comment';
  /** Author of the video comment */
  author: Membership;
  /** Timestamp of the block the comment was created at */
  createdAt: Scalars['DateTime'];
  /** METAPROTOCOL-{network}-{blockNumber}-{indexInBlock} */
  id: Scalars['String'];
  /** Whether comment has been edited or not */
  isEdited: Scalars['Boolean'];
  /** Whether a comment has been excluded/hidden (by the gateway operator) */
  isExcluded: Scalars['Boolean'];
  /** A (parent) comment that this comment replies to (if any) */
  parentComment?: Maybe<Comment>;
  /** List of all reactions to the comment */
  reactions: Array<CommentReaction>;
  /** Sum of replies and reactions */
  reactionsAndRepliesCount: Scalars['Int'];
  /** Total number of reactions to this comment */
  reactionsCount: Scalars['Int'];
  /** Reactions count by reaction Id */
  reactionsCountByReactionId?: Maybe<Array<CommentReactionsCountByReactionId>>;
  /** How many comments has replied to this comment */
  repliesCount: Scalars['Int'];
  /** Status of the comment; either it is visible, deleted, or moderated (deleted by moderator) */
  status: CommentStatus;
  /** Comment text */
  text: Scalars['String'];
  /** Video the comment was added to */
  video: Video;
};
// ...
export type DistributionBucket = {
  __typename?: 'DistributionBucket';
  /** Whether the bucket is accepting any new bags */
  acceptingNewBags: Scalars['Boolean'];
  /** Storage bags assigned to the bucket */
  bags: Array<DistributionBucketBag>;
  /** Bucket index within the family */
  bucketIndex: Scalars['Int'];
  /** Whether the bucket is currently distributing content */
  distributing: Scalars['Boolean'];
  /** Distribution family the bucket is part of */
  family: DistributionBucketFamily;
  /** Runtime bucket id in {familyId}:{bucketIndex} format */
  id: Scalars['String'];
  /** Distribution bucket operators (either active or invited) */
  operators: Array<DistributionBucketOperator>;
};
// ...
export type DistributionBucketBag = {
  __typename?: 'DistributionBucketBag';
  bag: StorageBag;
  distributionBucket: DistributionBucket;
  /** {distributionBucketId}-{storageBagId} */
  id: Scalars['String'];
};
// ...
export type DistributionBucketOperator = {
  __typename?: 'DistributionBucketOperator';
  /** Related distirbution bucket */
  distributionBucket: DistributionBucket;
  /** {bucketId}-{workerId} */
  id: Scalars['String'];
  /** Operator metadata */
  metadata?: Maybe<DistributionBucketOperatorMetadata>;
  /** Current operator status */
  status: DistributionBucketOperatorStatus;
  /** ID of the distribution group worker */
  workerId: Scalars['Int'];
};

export type DistributionBucketOperatorEdge = {
  __typename?: 'DistributionBucketOperatorEdge';
  cursor: Scalars['String'];
  node: DistributionBucketOperator;
};

export type DistributionBucketOperatorMetadata = {
  __typename?: 'DistributionBucketOperatorMetadata';
  /** Distribution bucket operator */
  distirbutionBucketOperator: DistributionBucketOperator;
  /** Additional information about the node/operator */
  extra?: Maybe<Scalars['String']>;
  id: Scalars['String'];
  /** Root distributor node api endpoint */
  nodeEndpoint?: Maybe<Scalars['String']>;
  /** Optional node location metadata */
  nodeLocation?: Maybe<NodeLocationMetadata>;
};
// ...
export type License = {
  __typename?: 'License';
  /** Attribution (if required by the license) */
  attribution?: Maybe<Scalars['String']>;
  /** License code defined by Joystream */
  code?: Maybe<Scalars['Int']>;
  /** Custom license content */
  customText?: Maybe<Scalars['String']>;
  /** Unique identifier */
  id: Scalars['String'];
};
// ...
export type Query = {
// ...
  videos: Array<Video>;
  videosConnection: VideosConnection;
};
// ...
export type StorageBag = {
  __typename?: 'StorageBag';
  /** Distribution buckets assigned to the bag */
  distributionBuckets: Array<DistributionBucketBag>;
  /** Storage bag id */
  id: Scalars['String'];
  /** Data objects in the bag */
  objects: Array<StorageDataObject>;
  /** Owner of the storage bag */
  owner: StorageBagOwner;
  /** Storage buckets assigned to the bag */
  storageBuckets: Array<StorageBucketBag>;
};
// ...
export type StorageDataObject = {
  __typename?: 'StorageDataObject';
  /** Timestamp of the block the data object was created at */
  createdAt: Scalars['DateTime'];
  /** Data object runtime id */
  id: Scalars['String'];
  /** IPFS content hash */
  ipfsHash: Scalars['String'];
  /** Whether the data object was uploaded and accepted by the storage provider */
  isAccepted: Scalars['Boolean'];
  /** Resolved asset urls */
  resolvedUrls: Array<Scalars['String']>;
  /** Data object size in bytes */
  size: Scalars['BigInt'];
  /** State Bloat Bond for removing the data object */
  stateBloatBond: Scalars['BigInt'];
  /** Storage bag the data object is part of */
  storageBag: StorageBag;
  /** The type of the asset that the data object represents (if known) */
  type?: Maybe<DataObjectType>;
  /** If the object is no longer used as an asset - the time at which it was unset (if known) */
  unsetAt?: Maybe<Scalars['DateTime']>;
};
// ...
export type Video = {
  __typename?: 'Video';
  /** Reference to a video category */
  category?: Maybe<VideoCategory>;
  /** Reference to videos's channel */
  channel: Channel;
  /** List of all video comments */
  comments: Array<Comment>;
  /** Comments count */
  commentsCount: Scalars['Int'];
  /** Timestamp of the block the video was created at */
  createdAt: Scalars['DateTime'];
  /** Block the video was created in */
  createdInBlock: Scalars['Int'];
  /** The description of the Video */
  description?: Maybe<Scalars['String']>;
  /** Video duration in seconds */
  duration?: Maybe<Scalars['Int']>;
  /** Application used for video creation */
  entryApp?: Maybe<App>;
  /** Whether or not Video contains marketing */
  hasMarketing?: Maybe<Scalars['Boolean']>;
  /** Runtime identifier */
  id: Scalars['String'];
  /** Flag signaling whether a video is censored. */
  isCensored: Scalars['Boolean'];
  /** Is comment section enabled (true if enabled) */
  isCommentSectionEnabled: Scalars['Boolean'];
  /** Whether a video has been excluded/hidden (by the gateway operator) */
  isExcluded: Scalars['Boolean'];
  /** Whether the Video contains explicit material. */
  isExplicit?: Maybe<Scalars['Boolean']>;
  /** Whether the Video is supposed to be publically displayed */
  isPublic?: Maybe<Scalars['Boolean']>;
  /** Is reactions feature enabled on video (true if enabled i.e. video can be reacted) */
  isReactionFeatureEnabled: Scalars['Boolean'];
  /** Video's main langauge */
  language?: Maybe<Scalars['String']>;
  /** License under the video is published */
  license?: Maybe<License>;
  /** Video media asset */
  media?: Maybe<StorageDataObject>;
  /** Video file metadata */
  mediaMetadata?: Maybe<VideoMediaMetadata>;
  /** Video NFT details */
  nft?: Maybe<OwnedNft>;
  /** channel owner pinned comment */
  pinnedComment?: Maybe<Comment>;
  /** If the Video was published on other platform before beeing published on Joystream - the original publication date */
  publishedBeforeJoystream?: Maybe<Scalars['DateTime']>;
  /** List of all video reactions */
  reactions: Array<VideoReaction>;
  /** Reactions count */
  reactionsCount: Scalars['Int'];
  /** Reactions count by reaction Id */
  reactionsCountByReactionId?: Maybe<Array<VideoReactionsCountByReactionType>>;
  /** List of video subtitles */
  subtitles: Array<VideoSubtitle>;
  /** Video thumbnail asset (recommended ratio: 16:9) */
  thumbnailPhoto?: Maybe<StorageDataObject>;
  /** The title of the video */
  title?: Maybe<Scalars['String']>;
  /** Value of video state bloat bond fee paid by channel owner */
  videoStateBloatBond: Scalars['BigInt'];
  /** Number of video views (to speed up orderBy queries by avoiding COUNT aggregation) */
  viewsNum: Scalars['Int'];
  /** Video ID coming from YPP */
  ytVideoId?: Maybe<Scalars['String']>;
};
// ...
export type VideoCategory = {
  __typename?: 'VideoCategory';
  createdInBlock: Scalars['Int'];
  /** The description of the category */
  description?: Maybe<Scalars['String']>;
  featuredVideos: Array<VideoFeaturedInCategory>;
  /** Runtime identifier */
  id: Scalars['String'];
  /** Indicates whether the category is supported by the Gateway */
  isSupported: Scalars['Boolean'];
  /** The name of the category */
  name?: Maybe<Scalars['String']>;
  /** Parent category if defined */
  parentCategory?: Maybe<VideoCategory>;
  videos: Array<Video>;
};
// ...
export type VideoReactionsCountByReactionType = {
  __typename?: 'VideoReactionsCountByReactionType';
  /** No of times the video has been reacted with given reaction */
  count: Scalars['Int'];
  /** The reaction option */
  reaction: VideoReactionOptions;
};
// ...
export type GetVideosQuery = { __typename?: 'Query', videos: Array<{ __typename?: 'Video', id: string, media?: (
      { __typename?: 'StorageDataObject' }
      & { ' $fragmentRefs'?: { 'StorageDataObjectFieldsFragment': StorageDataObjectFieldsFragment } }
    ) | null, thumbnailPhoto?: (
      { __typename?: 'StorageDataObject' }
      & { ' $fragmentRefs'?: { 'StorageDataObjectFieldsFragment': StorageDataObjectFieldsFragment } }
    ) | null }> };

export type StorageDataObjectFieldsFragment = { __typename?: 'StorageDataObject', id: string, storageBag: { __typename?: 'StorageBag', distributionBuckets: Array<{ __typename?: 'DistributionBucketBag', distributionBucket: { __typename?: 'DistributionBucket', id: string, operators: Array<{ __typename?: 'DistributionBucketOperator', workerId: number, metadata?: { __typename?: 'DistributionBucketOperatorMetadata', nodeEndpoint?: string | null } | null }> } }> } } & { ' $fragmentName'?: 'StorageDataObjectFieldsFragment' };